agentiq/mvp/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Celery for background tasks
celery==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Telegram
python-telegram-bot==20.8

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM (DeepSeek via OpenAI SDK)
openai>=1.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
